git-svn-id: https://svn.st.cs.uni-sb.de/svn/st/projects/java-slicer/trunk@7226 ef2ef919-bb9b-4b18-8ff0-e5cc2bf85462

+++ javaslicer-traceReader/src/main/java/de/unisb/cs/st/javaslicer/traceResult/traceSequences/ConstantUncompressedLongTraceSequence.java
@@ -68,7 +68,7 @@ public class ConstantUncompressedLongTraceSequence implements ConstantLongTraceS
                     return true;
                 if (this.offset == 0)
                     return false;
-                this.offset -= this.buf.length*8;
+                this.offset -= (long)this.buf.length*8;
                 this.inputStream.seek(this.offset);
                 for (int i = 0; i < this.buf.length; ++i) {
                     this.buf[i] = this.dataIn.readLong();
